authy: move authy calls to pam_authy
