+++ src/pam_util.c
@@ -1,0 +1,72 @@
+#include "config.h"
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <stdarg.h>
+#include <unistd.h>
+#include <string.h>
+#include <syslog.h>
+
+#include <security/pam_appl.h>
+#include <security/pam_modules.h>
+
+#include "pam_util.h"
+
+#if defined(DEMO) || defined(TESTING)
+static char* error_msg = NULL;
+
+const char *get_error_msg(void) __attribute__((visibility("default")));
+const char *get_error_msg(void)
+{
+  if (!error_msg) {
+    return "";
+  }
+  return error_msg;
+}
+#endif
+
+void log_message(int priority, pam_handle_t *pamh,
+                        const char *format, ...)
+{
+  char *service = NULL;
+  if (pamh)
+    pam_get_item(pamh, PAM_SERVICE, (void *)&service);
+  if (!service)
+    service = "";
+
+  char logname[80];
+  snprintf(logname, sizeof(logname), "%s(" MODULE_NAME ")", service);
+
+  va_list args;
+  va_start(args, format);
+#if !defined(DEMO) && !defined(TESTING)
+  openlog(logname, LOG_CONS | LOG_PID, LOG_AUTHPRIV);
+  vsyslog(priority, format, args);
+  closelog();
+#else
+  if (!error_msg) {
+    error_msg = strdup("");
+  }
+  {
+    char buf[1000];
+    vsnprintf(buf, sizeof buf, format, args);
+    const int newlen = strlen(error_msg) + 1 + strlen(buf) + 1;
+    char* n = malloc(newlen);
+    if (n) {
+      snprintf(n, newlen, "%s%s%s", error_msg, strlen(error_msg)?"\n":"",buf);
+      free(error_msg);
+      error_msg = n;
+    } else {
+      fprintf(stderr, "Failed to malloc %d bytes for log data.\n", newlen);
+    }
+  }
+#endif
+
+  va_end(args);
+
+  if (priority == LOG_EMERG) {
+    // Something really bad happened. There is no way we can proceed safely.
+    _exit(1);
+  }
+}
+
